Change temporarystat option to int

+++ src/Edelstein.Service.Game/Fields/User/Stats/TemporaryStat.cs
@@ -6,7 +6,7 @@ namespace Edelstein.Service.Game.Fields.User.Stats
     {
         public TemporaryStatType Type { get; set; }
 
-        public short Option { get; set; }
+        public int Option { get; set; }
         public int TemplateID { get; set; }
         public DateTime? DateExpire { get; set; }
     }
